requirements/prod.txt
fastapi[standard]==0.115.4
pydantic-settings==2.6.0
sqlalchemy==2.0.35
aiohttp==3.10.10
asyncpg==0.30.0
telethon==1.37.0
apscheduler==3.10.4
numpy==1.26.4
scikit-learn==1.5.2
langchain-community==0.3.5
gigachat==0.1.35
faiss-cpu
